Move BaseEngine to root and remove log_train/log_eval methods
