--- docs/source/engine.rst
@@ -1,8 +1,0 @@
-lagom.engine
-===================================
-
-.. automodule:: lagom.engine
-.. currentmodule:: lagom.engine
-
-.. autoclass:: BaseEngine
-    :members:
