+++ docs/source/index.rst
@@ -50,7 +50,6 @@ A graphical illustration is coming soon.
     lagom
     lagom.agents <agents>
     lagom.contrib <contrib>
-    lagom.engine <engine>
     lagom.envs <envs>
     lagom.es <es>
     lagom.experiment <experiment>
