+++ docs/source/lagom.rst
@@ -6,3 +6,8 @@ Logger
 ----------------------------
 .. autoclass:: Logger
     :members:
+    
+Engine
+----------------------------
+.. autoclass:: BaseEngine
+    :members:
