+++ lagom/__init__.py
@@ -1,3 +1,5 @@
 from .version import __version__
 
 from .logger import Logger
+
+from .base_engine import BaseEngine
