--- lagom/engine/__init__.py
@@ -1,1 +1,0 @@
-from .base_engine import BaseEngine
